moved user-routes.js inside the api directory, removed api versioning in DocumentClient()
