+++ server/server.js
@@ -1,27 +1,26 @@
-const express = require('express');
+const express = require("express");
 const app = express();
 const PORT = process.env.PORT || 3001;
-const fileRoutes = require('./routes/file-upload');
-const userRoutes = require('./routes/user-routes');
+const fileRoutes = require("./routes/file-upload");
+const userRoutes = require("./routes/api/user-routes");
 
 // express middleware, used to be bodyparser
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
 // Serve up static assets
-if (process.env.NODE_ENV === 'production') {
-  app.use(express.static('client/build'));
+if (process.env.NODE_ENV === "production") {
+  app.use(express.static("client/build"));
 }
 
 // app.use(require('./routes'));
-app.use('/api/', fileRoutes);
-app.use('/api/', userRoutes);
+app.use("/api/", fileRoutes);
+app.use("/api/", userRoutes);
 
 // Start the API server
 app.listen(PORT, () =>
   console.log(`🌎  ==> API Server now listening on PORT ${PORT}!`)
 );
-
 
 // const express = require('express');
 // const path = require('path');
